HP-292 Extract common authentication test stuff into their own file
